requirements.txt
asgiref==3.7.2
Django==5.0
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
